+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.6.8
 
- - mssql: Add support for DATETIME (in addition to the existing DATETIME2)
+ - mssql: Add support for DATETIME and SMALLDATETIME (in addition to the existing DATETIME2)
 
 ## 0.6.7